Face detection/requirements .txt
opencv-python
pyserial
tensorflow
numpy